:bug: Builder - Fix caching issue in builder
