+++ client/src/builder/hooks/use-export-modal-store.ts
@@ -1,0 +1,11 @@
+import { create } from "zustand";
+
+type ExportModalStore = {
+  isOpen: boolean;
+  setOpen: (open: boolean) => void;
+};
+
+export const useExportModalStore = create<ExportModalStore>((set) => ({
+  isOpen: false,
+  setOpen: (isOpen) => set({ isOpen }),
+}));
